chore: use BigDecimal instead of double for division

+++ src/main/java/school/hei/haapi/service/GradeResultService.java
@@ -26,7 +26,7 @@ public class GradeResultService {
 
     return new YearlyResult()
         .level(level)
-        .weightedAverage(courseResultUtils.weightedSumOfCourseResults(courseResults))
+        .weightedAverage(courseResultUtils.weightedSumOfCourseResults(courseResults).doubleValue())
         .obtainedCredits(courseResultUtils.obtainedCreditsOfCourseResults(courseResults))
         .courseResults(courseResults);
   }
